// app/core/hexag/documentos/domain/enums/tipo-documento.enum.ts
/**
 * Tipos de documentos que se pueden generar para una junta
 */
export enum TipoDocumento {
  // Acta
  ACTA = "ACTA",
  FALTA_QUORUM = "FALTA_QUORUM",

  // Documentos No-Punto
  CONVOCATORIA = "CONVOCATORIA",
  PROXY_NATURAL = "PROXY_NATURAL",
  PROXY_JURIDICA = "PROXY_JURIDICA",
  CERTIFICACION = "CERTIFICACION",
  LISTA_ASISTENCIA = "LISTA_ASISTENCIA",

  // Documentos por Punto - Aporte Dinerario
  MINUTA_APORTE_DINERARIO = "MINUTA_APORTE_DINERARIO",
  ASIENTO_APORTE_DINERARIO = "ASIENTO_APORTE_DINERARIO",
  CERTIFICADO_APORTE_DINERARIO = "CERTIFICADO_APORTE_DINERARIO",
  AVISO_APORTE_DINERARIO = "AVISO_APORTE_DINERARIO",
  CARTA_AVISO_APORTE_DINERARIO = "CARTA_AVISO_APORTE_DINERARIO",

  // Documentos por Punto - Capitalización
  MINUTA_CAPITALIZACION = "MINUTA_CAPITALIZACION",
  ASIENTO_CAPITALIZACION = "ASIENTO_CAPITALIZACION",
  CERTIFICADO_CAPITALIZACION = "CERTIFICADO_CAPITALIZACION",
  INFORME_CREDITOS = "INFORME_CREDITOS",

  // Documentos por Punto - Nombramiento
  CERTIFICACION_2_NOMBRAMIENTO = "CERTIFICACION_2_NOMBRAMIENTO",
  SOLICITUD_COPIAS = "SOLICITUD_COPIAS",
  ACEPTACION_CARGO = "ACEPTACION_CARGO",
  RENUNCIA = "RENUNCIA",
}